Renames all protocol references to handlers.
