+++ extensions/patient-visit-summary/patient_visit_summary/CANVAS_MANIFEST.json
@@ -1,12 +1,12 @@
 {
     "sdk_version": "0.134.0",
-    "plugin_version": "0.0.69",
+    "plugin_version": "0.0.70",
     "name": "patient_visit_summary",
     "description": "Generate a visit summary printout or customizable print view for a patient note.",
     "components": {
-        "protocols": [
+        "handlers": [
             {
-                "class": "patient_visit_summary.protocols.patient_visit_summary:PatientVisitSummaryButton",
+                "class": "patient_visit_summary.handlers.patient_visit_summary:PatientVisitSummaryButton",
                 "description": "Note header button that launches the Patient Visit Summary modal.",
                 "data_access": {
                     "event": "",
@@ -15,7 +15,7 @@
                 }
             },
             {
-                "class": "patient_visit_summary.protocols.patient_visit_summary:PatientVisitSummaryAPI",
+                "class": "patient_visit_summary.handlers.patient_visit_summary:PatientVisitSummaryAPI",
                 "description": "API endpoint for the Patient Visit Summary HTML.",
                 "data_access": {
                     "event": "",
@@ -24,7 +24,7 @@
                 }
             },
             {
-                "class": "patient_visit_summary.protocols.customize_print:CustomizePrintButton",
+                "class": "patient_visit_summary.handlers.customize_print:CustomizePrintButton",
                 "description": "Note footer button that launches the Customize & Print modal.",
                 "data_access": {
                     "event": "",
@@ -33,7 +33,7 @@
                 }
             },
             {
-                "class": "patient_visit_summary.protocols.customize_print:CustomizePrintAPI",
+                "class": "patient_visit_summary.handlers.customize_print:CustomizePrintAPI",
                 "description": "API endpoint for the Customize & Print modal HTML.",
                 "data_access": {
                     "event": "",
